El atributo num_pages se ha acotado.
Un campo IntegerField además de ser más voluminoso de lo que se busca,
da la posibilidad de introducir números negativos lo que no tendría
sentido para el atributo.

+++ bookstore_inventory/models/book.py
@@ -6,7 +6,7 @@ class Book(models.Model):
     author    = models.CharField(max_length = 100)
     editorial = models.CharField(max_length = 150)
     gender    = models.CharField(max_length = 100)
-    num_pages = models.IntegerField(verbose_name = "number of pages")
+    num_pages = models.PositiveSmallIntegerField(verbose_name = "number of pages")
     pub_date  = models.DateField(verbose_name = "publication date")
 
     class Meta:
